ストラテジーのリファクタ
